Reader and Writer for Sudoku now work as advertized
You parse a problem description with a Reader. The Reader can
then provide a matrix for a Solver to solve, as well as a Writer
to output each solution that the Solver finds.

The matrix generated by the Reader is pre-compacted by having
the rows corresponding to hints in the grid removed, and then
re-numbeing all the columns so that there are no gaps. That is,
all the row and column numbers are dependent not just on the
grid size, but also on the hints in the grid.

You can generate a Writer on your own if you want, but it will
need a mapping translating matrix rows into row/column/number.
The Reader does this for you, creating a Writer from its internal
mapping.