[c] modified to count solutions, not print them
The C implementation stored all the solutions in a big
data structure and printed them at the end. This is not
viable if the number of solutions go into the trillions.
So here's an alternative; for each solution, increment
a variable, and print its value at the end. Also print
the value of the variable with every millionth increment.